Procon
Producer-Consumer (ProCon) is a classic problem in programming where one or more producers create data and put it into a buffer, and one or more consumers take data from the buffer to process it. The challenge is to coordinate them so producers don’t overflow the buffer and consumers don’t try to consume empty data. It’s used to manage synchronization in concurrent systems.